+++ plugins/lookup/aws_account_attribute.py
@@ -52,6 +52,8 @@ _raw:
     (or all attributes if one is not specified).
 """
 
+import os
+
 from ansible.errors import AnsibleError
 
 try:
@@ -60,12 +62,13 @@ try:
 except ImportError:
     raise AnsibleError("The lookup aws_account_attribute requires boto3 and botocore.")
 
-from ansible.plugins import AnsiblePlugin
-from ansible.plugins.lookup import LookupBase
-from ansible_collections.amazon.aws.plugins.module_utils.ec2 import boto3_conn, get_aws_connection_info
 from ansible.module_utils._text import to_native
 from ansible.module_utils.six import string_types
-import os
+from ansible.plugins import AnsiblePlugin
+from ansible.plugins.lookup import LookupBase
+
+from ..module_utils.ec2 import boto3_conn
+from ..module_utils.ec2 import get_aws_connection_info
 
 
 def _boto3_conn(region, credentials):
