+++ plugins/module_utils/cloudfront_facts.py
@@ -29,13 +29,16 @@ Common cloudfront facts shared between modules
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
-from ansible_collections.amazon.aws.plugins.module_utils.ec2 import get_aws_connection_info, boto3_conn
-from ansible_collections.amazon.aws.plugins.module_utils.ec2 import boto3_tag_list_to_ansible_dict, camel_dict_to_snake_dict
-
 try:
     import botocore
 except ImportError:
     pass
+
+from ansible.module_utils.common.dict_transformations import camel_dict_to_snake_dict
+
+from .ec2 import boto3_conn
+from .ec2 import boto3_tag_list_to_ansible_dict
+from .ec2 import get_aws_connection_info
 
 
 class CloudFrontFactsServiceManager(object):
