+++ plugins/module_utils/direct_connect.py
@@ -33,11 +33,13 @@ from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
 import traceback
+
 try:
     import botocore
 except ImportError:
     pass
-from ansible_collections.amazon.aws.plugins.module_utils.ec2 import camel_dict_to_snake_dict
+
+from ansible.module_utils.common.dict_transformations import camel_dict_to_snake_dict
 
 
 class DirectConnectError(Exception):
