+++ plugins/module_utils/ec2.py
@@ -34,15 +34,20 @@ import re
 import sys
 import traceback
 
+from ansible.module_utils._text import to_native
+from ansible.module_utils._text import to_text
 from ansible.module_utils.ansible_release import __version__
-from ansible.module_utils.basic import missing_required_lib, env_fallback
-from ansible.module_utils._text import to_native, to_text
-from ansible_collections.amazon.aws.plugins.module_utils.cloud import CloudRetry
-from ansible.module_utils.six import string_types, binary_type, text_type
-from ansible.module_utils.common.dict_transformations import (
-    camel_dict_to_snake_dict, snake_dict_to_camel_dict,
-    _camel_to_snake, _snake_to_camel,
-)
+from ansible.module_utils.basic import env_fallback
+from ansible.module_utils.basic import missing_required_lib
+from ansible.module_utils.common.dict_transformations import _camel_to_snake
+from ansible.module_utils.common.dict_transformations import _snake_to_camel
+from ansible.module_utils.common.dict_transformations import camel_dict_to_snake_dict
+from ansible.module_utils.common.dict_transformations import snake_dict_to_camel_dict
+from ansible.module_utils.six import binary_type
+from ansible.module_utils.six import string_types
+from ansible.module_utils.six import text_type
+
+from .cloud import CloudRetry
 
 BOTO_IMP_ERR = None
 try:
