+++ plugins/module_utils/waf.py
@@ -32,13 +32,15 @@ This module adds shared support for Web Application Firewall modules
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
-from ansible_collections.amazon.aws.plugins.module_utils.ec2 import camel_dict_to_snake_dict, AWSRetry
-from ansible_collections.amazon.aws.plugins.module_utils.aws.waiters import get_waiter
-
 try:
     import botocore
 except ImportError:
     pass  # caught by imported HAS_BOTO3
+
+from ansible.module_utils.common.dict_transformations import camel_dict_to_snake_dict
+
+from .ec2 import AWSRetry
+from .waiters import get_waiter
 
 
 MATCH_LOOKUP = {
