+++ plugins/modules/cloudformation.py
@@ -323,8 +323,8 @@ stack_outputs:
 
 import json
 import time
-import uuid
 import traceback
+import uuid
 from hashlib import sha1
 
 try:
@@ -335,10 +335,11 @@ except ImportError:
 
 from ansible.module_utils._text import to_bytes
 from ansible.module_utils._text import to_native
-from ansible_collections.amazon.aws.plugins.module_utils.aws.core import AnsibleAWSModule
-from ansible_collections.amazon.aws.plugins.module_utils.ec2 import ansible_dict_to_boto3_tag_list
-from ansible_collections.amazon.aws.plugins.module_utils.ec2 import AWSRetry
-from ansible_collections.amazon.aws.plugins.module_utils.ec2 import boto_exception
+
+from ..module_utils.core import AnsibleAWSModule
+from ..module_utils.ec2 import AWSRetry
+from ..module_utils.ec2 import ansible_dict_to_boto3_tag_list
+from ..module_utils.ec2 import boto_exception
 
 
 def get_stack_events(cfn, stack_name, events_limit, token_filter=None):
