+++ plugins/modules/ec2.py
@@ -599,10 +599,11 @@ from ansible.module_utils.six import get_function_code
 from ansible.module_utils.six import string_types
 from ansible.module_utils._text import to_bytes
 from ansible.module_utils._text import to_text
-from ansible_collections.amazon.aws.plugins.module_utils.aws.core import AnsibleAWSModule
-from ansible_collections.amazon.aws.plugins.module_utils.ec2 import HAS_BOTO
-from ansible_collections.amazon.aws.plugins.module_utils.ec2 import ec2_connect
-from ansible_collections.amazon.aws.plugins.module_utils.ec2 import get_aws_connection_info
+
+from ..module_utils.core import AnsibleAWSModule
+from ..module_utils.ec2 import HAS_BOTO
+from ..module_utils.ec2 import ec2_connect
+from ..module_utils.ec2 import get_aws_connection_info
 
 
 def find_running_instances_by_count_tag(module, ec2, vpc, count_tag, zone=None):
